backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tflite-runtime==2.14.0
numpy<2
pillow
opencv-python